ml-backend/requirements.txt
fastapi==0.74.0
openai==0.11.0
python-dotenv==0.19.2
uvicorn==0.17.5
